Update 2D_params.f
testing - change day back to sp

+++ 2D_params.f
@@ -67,7 +67,7 @@
       !real(sp) :: time
       !real(sp) :: thour
       !real(sp) :: tmon
-      real(dp) :: day
+      real(sp) :: day
       !real(sp) :: dday
       !real(sp) :: hday
       !integer :: kday
